refactored codebase and docstrings

+++ protbench/embedder/utils.py
@@ -1,10 +1,13 @@
+from __future__ import annotations
+
 import os
 from dataclasses import dataclass
-from functools import partial
+from os import PathLike
+from pathlib import Path
 from typing import Callable
 from typing import Dict
 from typing import List
-from typing import Optional
+from typing import NamedTuple
 
 import torch
 
@@ -12,47 +15,56 @@ from protbench.embedder import TorchEmbedder
 from protbench.embedder import TorchEmbeddingFunction
 
 
+def delete_directory_contents(directory: PathLike):
+    if not Path(directory).exists():
+        return
+    for root, _, files in os.walk(directory, topdown=False):
+        for name in files:
+            os.remove(os.path.join(root, name))
+
+
 @dataclass
 class SaveDirectories:
-    parent_dir: str = "/root/.cache"
-    train_dir: str = "train_embeddings"
-    validation_dir: str = "validation_embeddings"
-    test_dir: str = "test_embeddings"
+    parent_dir: PathLike = "/root/.cache"
+    train_dir: PathLike = "train_embeddings"
+    validation_dir: PathLike = "validation_embeddings"
+    test_dir: PathLike = "test_embeddings"
+
+    def create_or_clean_directory(self, directory: PathLike):
+        delete_directory_contents(directory=directory)
+        Path(directory).mkdir(exist_ok=True)
 
     @property
-    def train(self):
+    def train(self) -> PathLike:
         return os.path.join(self.parent_dir, self.train_dir)
 
     @property
-    def validation(self):
+    def validation(self) -> PathLike:
         return os.path.join(self.parent_dir, self.validation_dir)
 
     @property
-    def test(self):
+    def test(self) -> PathLike:
         return os.path.join(self.parent_dir, self.test_dir)
-
-
-def delete_directory_contents(directory_path: str):
-    for root, _, files in os.walk(directory_path, topdown=False):
-        for name in files:
-            os.remove(os.path.join(root, name))
 
 
 def compute_embeddings(
     model: torch.nn.Module,
     tokenizer: Callable,
     train_seqs: List[str],
-    val_seqs: Optional[List[str]] = None,
-    test_seqs: Optional[List[str]] = None,
+    val_seqs: List[str] | None = None,
+    test_seqs: List[str] | None = None,
+    forward_options: Dict | None = {},
     post_processing_fn: Callable = None,
-    device: Optional[torch.device] = None,
+    device: torch.device | None = None,
     pad_token_id: int = 0,
     tokenizer_options: Dict = {},
 ):
     embedding_fn = TorchEmbeddingFunction(
-        model,
-        partial(tokenizer, **tokenizer_options),
+        model=model,
+        tokenizer=tokenizer,
         device=device,
+        tokenizer_options=tokenizer_options,
+        forward_options=forward_options,
         embeddings_postprocessing_fn=post_processing_fn,
         pad_token_id=pad_token_id,
     )
@@ -65,15 +77,6 @@ def compute_embeddings(
     )
 
     embeddings = []
-    compute_data = [
-        train_seqs,
-    ]
-
-    if val_seqs is not None:
-        compute_data.append(val_seqs)
-
-    if test_seqs is not None:
-        compute_data.append(test_seqs)
 
     for data in [train_seqs, val_seqs, test_seqs]:
         if data is not None:
@@ -89,46 +92,36 @@ def compute_embeddings_and_save_to_disk(
     tokenizer: Callable,
     save_directories: SaveDirectories,
     train_seqs: List[str],
-    val_seqs: Optional[List[str]] = None,
-    test_seqs: Optional[List[str]] = None,
+    val_seqs: List[str] | None = None,
+    test_seqs: List[str] | None = None,
+    forward_options: Dict | None = {},
     post_processing_fn: Callable = None,
-    device: Optional[torch.device] = None,
+    device: torch.device | None = None,
     pad_token_id: int = 0,
     tokenizer_options: Dict = {},
 ):
     embedding_fn = TorchEmbeddingFunction(
-        model,
-        partial(tokenizer, **tokenizer_options),
+        model=model,
+        tokenizer=tokenizer,
         device=device,
+        tokenizer_options=tokenizer_options,
+        forward_options=forward_options,
         embeddings_postprocessing_fn=post_processing_fn,
         pad_token_id=pad_token_id,
     )
 
-    if not os.path.exists(save_directories.train):
-        os.mkdir(save_directories.train)
-    else:
-        delete_directory_contents(save_directories.train)
-
-    if val_seqs is not None and not os.path.exists(
-        save_directories.validation
-    ):
-        os.mkdir(save_directories.validation)
-    else:
-        delete_directory_contents(save_directories.validation)
-
-    if test_seqs is not None and not os.path.exists(save_directories.test):
-        os.mkdir(save_directories.test)
-    else:
-        delete_directory_contents(save_directories.test)
+    save_directories.create_or_clean_directory(save_directories.train)
 
     compute_data = [
         (train_seqs, save_directories.train),
     ]
 
     if val_seqs is not None:
+        save_directories.create_or_clean_directory(save_directories.validation)
         compute_data.append((val_seqs, save_directories.validation))
 
     if test_seqs is not None:
+        save_directories.create_or_clean_directory(save_directories.test)
         compute_data.append((test_seqs, save_directories.test))
 
     for data, path in compute_data:
@@ -142,11 +135,10 @@ def compute_embeddings_and_save_to_disk(
         embedder.run(data)
 
 
-@dataclass
-class EmbeddingsContainer:
+class EmbeddingsContainer(NamedTuple):
     train_embeddings: List[torch.Tensor]
-    val_embeddings: Optional[List[torch.Tensor]] = None
-    test_embeddings: Optional[List[torch.Tensor]] = None
+    val_embeddings: List[torch.Tensor] | None = None
+    test_embeddings: List[torch.Tensor] | None = None
 
 
 class ComputeEmbeddingsWrapper:
@@ -155,11 +147,12 @@ class ComputeEmbeddingsWrapper:
         model: torch.nn.Module,
         tokenizer: Callable,
         tokenizer_options: Dict = {},
-        post_processing_function: Optional[Callable] = None,
+        forward_options: Dict = {},
+        post_processing_function: Callable | None = None,
         device: torch.device = None,
         pad_token_id: int = 0,
         low_memory: bool = False,
-        save_directories: Optional[SaveDirectories] = None,
+        save_directories: SaveDirectories | None = None,
     ):
         if low_memory and save_directories is None:
             raise ValueError(
@@ -179,10 +172,11 @@ class ComputeEmbeddingsWrapper:
         self.pad_token_id = pad_token_id
         self.low_memory = low_memory
         self.save_directories = save_directories
+        self.forward_options = forward_options
 
     def __call__(
         self, train_seqs, val_seqs=None, test_seqs=None
-    ) -> EmbeddingsContainer:
+    ) -> EmbeddingsContainer | None:
         if self.low_memory:
             compute_embeddings_and_save_to_disk(
                 self.model,
